+++ ecaudit/src/main/java/com/ericsson/bss/cassandra/ecaudit/auth/AuditAuthorizer.java
@@ -134,4 +134,9 @@ public class AuditAuthorizer implements IAuthorizer
     {
         wrappedAuthorizer.setup();
     }
+
+    public IAuthorizer getWrappedAuthorizer()
+    {
+        return wrappedAuthorizer;
+    }
 }